Rework authentication methods, add ctauth-based method
This commit renames httpauth to gitauth and adds a new auth
mechanism (codethinkauth, based on python-ctauth). It also
moves common functionality (the whole function decoration and
wrapping to protect pages) into a common base class called
mustard.auth.Authenticator.

In addition to this, the following command line options are
introduced:

  --auth=<codethink|git> (default: codethink)
  --auth-server=...      (e.g. ldap://your.ldap.server.org/)
  --auth-user=...        (e.g. LDAP lookup user)
  --auth-password=...    (e.g. LDAP lookup user password)
  --project-code=XYZ