Merge architectures and components, render them hierarchically
This decision was made because the separation of architectures and
components was felt to be artificial and got in the way of defining
hierarchies.

This commit moves all fields from architectures into components and
makes "kind: architecture", "kind: arch", "kind: a" aliases for
"kind: component". This provides backwards-compatibility and allows
to browse MUSTARD repos that haven't removed architectures yet.

This commit also introduces hierarchical rendering of components
(and architectures thanks to the introduced aliases) in the same
way requirement and work item hierarchies are rendered.

Signed-off-by: Jannis Pohlmann <jannis.pohlmann@codethink.co.uk>